// src/types/Product.types.ts
export interface ProductProps {
  id: string
  name: string
  imageUrl: string
  description: string
  price: number
  priceFormatted: string
  defaultPriceId: string
}

export interface StorageProductProps {
  id: string
  name: string
  imageUrl: string
  price: number
  priceFormatted: string
  defaultPriceId: string
}